Add dialer package. Put main into base directory. Add cobra commands
