+++ main.go
@@ -1,0 +1,32 @@
+// Copyright © 2019 KIM KeepInMind GmbH/srl
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see <http://www.gnu.org/licenses/>.
+
+package main
+
+import "github.com/booster-proj/booster/cmd"
+
+// Version and BuildTime are filled in during build by the Makefile
+var (
+	version   = "N/A"
+	commit    = "N/A"
+	buildTime = "N/A"
+)
+
+func main() {
+	cmd.Version = version
+	cmd.Commit = commit
+	cmd.BuildTime = buildTime
+	cmd.Execute()
+}
